added to clean.sh and added set_path method for Youtube2Mp3 class. still looking at the progress bar issue

+++ clean.sh
@@ -1,1 +1,7 @@
-rm -fr ~/Downloads/youtube2mp3/
+#!/usr/bin/bash
+if [[ -d ~/Downloads/youtube2mp3 ]];then
+    echo "output directory found, removing now"
+    rm -fr ~/Downloads/youtube2mp3/
+else
+    echo "--nothing to do for clean.sh"
+fi
